fix(availability): move availability to user
